requirements.txt
Django>=4.2.0,<4.3
djangorestframework>=3.14,<3.15
pre-commit==3.6.0
psycopg2>=2.9.3,<2.10
drf-spectacular==0.27.1
coverage==7.4.1
